restructure into separate directories for separate tkinter tools
