+++ diablo/detect_mana.h5
